Agrego api rest tipo DELETE para eliminar plan semanal por id

+++ src/main/java/ffleitas/back/service/PlanService.java
@@ -11,5 +11,5 @@ public interface PlanService {
     List<PlanSemanalDTO> listarPlanes();
     PlanSemanalDetalleDTO obtenerDetalleDePlanPorId(Long id);
     PlanSemanalDTO crearPlan(PlanSemanalRequest planSemanalRequest);
-    void eliminarPlanPorId(Long id);
+    void eliminarPlanPorId(Integer id);
 }
